// libs/core/src/lib/domain/user/entity/mock/user.mock.types.ts
export type IMergeUsernames =
  | 'idiot'
  | 'role-land-expert'
  | 'product-owner-team-a-b'
  | 'backend-developer-1-team-a'
  | 'backend-developer-2-team-a'
  | 'frontend-developer-1-team-a'
  | 'backend-developer-1-team-b'
  | 'backend-developer-2-team-b'
  | 'frontend-developer-1-team-b'
  | 'frontend-developer-2-team-b'
  | 'dev-ops-team-b'
  | 'qa-team-b'
  | 'client-team-b';